// src/contract-apis/types/OneClicks.ts
export declare namespace OneClicks {
  export type PositionsToCloseStruct = {
    underlier: PromiseOrValue<string>;
    positionIds: PromiseOrValue<BigNumberish>[];
  };

  export type PositionsToCloseStructOutput = [string, BigNumber[]] & {
    underlier: string;
    positionIds: BigNumber[];
  };

  export type LegDetailsStruct = {
    amount: PromiseOrValue<BigNumberish>;
    strikeId: PromiseOrValue<BigNumberish>;
    isCall: PromiseOrValue<boolean>;
    isLong: PromiseOrValue<boolean>;
    finalPositionCollateral: PromiseOrValue<BigNumberish>;
  };

  export type LegDetailsStructOutput = [
    BigNumber,
    BigNumber,
    boolean,
    boolean,
    BigNumber
  ] & {
    amount: BigNumber;
    strikeId: BigNumber;
    isCall: boolean;
    isLong: boolean;
    finalPositionCollateral: BigNumber;
  };
}